Fixed naive cross corr implementation
The original naive cross correlation implementation slides the reference
subregion over the deformed subregion. This is inconsistent with the
thesis text, where it is said to compute reference * deformed cross
correlation, which should slide the deformed over the reference.

Changed my implementation to slide the correct subregion. The new
implementation also matches the xcorr2 implementation from matlab so we
can validate the results.

Changed kernel signatures to have the ref and deformed arguments in the
correct order, based on the cross correlation we compute.

Changed to fixed precision output when writing matrix to CSV. This
allows us to keep the precision when importing the matrix to matlab.
Without this, values were written in scientific notation, losing some
precision.