First somewhat working version
Added fft based implementation. This implementation uses cuFFT to
compute FFT of the input matricies, zero padded to the required size
(twice the size of the original matrix). It then computes hadamard
product of complex conjugate of the FFT of one matrix with the FFT of
second matrix and then does IFFT to get circularly shifted results with
one additional row, as described in the original thesis.

There are few caveat in this FFT implementation. cuFFT does unnormalized
FFT, so IFFT(FFT(X)) has each element of the results multiplied by the
number of elements of X, so for 20x20 matrix X, the results at i,j is
the element at i,j from X multiplied by 400. When searching for the
position of maximum, the exact values do not matter, so we can ignore
this multiplication. But when validating the results with results
computed through matlab, we have to take this into account.

There is also a problem with the complex conjugate. Based on the
definition, it would seem taht the first (left) argument should be
conjugated, but this does not match the results from matlab xcorr2
function. For the results to match, we would need to conjugate the
second (right) argument. CUrrently we match the definition, but we
should discuss this further.

Another fun thing was to get the size of buffers and the size to pass
the cufft plans right. The documentation is not really that clear on
first readthrough, but it is there. The size to pass to the plan is the
size of the input matrix. The size of the required buffer is defined in
the documentation, and is equal to roughly half of the original matrix,
but with complex numbers, so almost the same size. The buffer size is
defined in the documentation as size().x * (floor(size().y / 2) + 1). As
our matricies are padded to double their size as part of the algorithm,
the floor operation can be ignored. So we are left with size().x *
(size().y / 2 + 1) of complex numbers.

cufft also has different coordinate system. They use x for rows and y
for columns, so y is the dimension changing faster than x. Whereas when
working with images, as in the original thesis, x is used for columns
and y for rows as if working with a monitor coordinates. As we are
working with abstract matricies, we could switch to the system used by
cufft and matlab, but that is a decision for the future. Currently the
only place this slightly matters is when passing size to cufft plans, as
we have to pass them inverted. But because we are currently working with
square matricies, it does not really matter.

Borrowed cufft helpers from the original thesis. I think it would be
waste of time to reimplement them myself, as they seem pretty much
logical wrappers around the cufft API.

Added cuFFT library to compilation. This required some workarounds as we
are using cmake 3.10, which does not really support CUDA that well.
Optimal way would be to use cmake 3.18, but Ubuntu 20.04 comes with
cmake 3.16, so we cannot reasonably expect anyone to have access to 3.18
yet.

Added some abstractions, such as cross_corr_alg base class for all
algorithms, to provide some shared functionality and allow us some type
safety when implementing algorithms.

Moved input file reading to the algorithms themselfs, so that the fft
algorithm can pad the matricies right when reading them.

Added a system to add padding when reading matricies, so that we can
allocate the whole buffer for the matrix and the padding in one go and
don't have to do it in multiple steps.

The padding implementation during reading is pretty simple, and just
allows us to append some values to the remainder of each row and then
add some full rows underneath, so that the original matrix is always in
the top left corner. This should be enough for our use case.

Added matlab implementation mirroring our C++ implementation, so that we
can validate the output of each step of the C++ implementation, while
also allowing us to validate the final output and reassemble the output
from the circularly shifted output of the ifft.