Added validation to C++ code
Added simple CPU cross correlation implementation which can be used to
validate the results of more advanced implementations.

Added FFT helper which reorders the resulting matrix, swapping quadrants
and removing the "empty" column and row. It also divides the results by
the number of elements of the resulting matrix, which solves the problem
with cuFFT computing unnormalized FFT.

Changed Hadamard implementation, switching which of the two values is
conjugated. This basically transposes the resulting matrix around the
minor diagonal. With this change, the results match the ones computed by
the simple CPU implementation, GPU naive implementation and matlab.

Added result validation to the C++ program itself, which now computes
statistics about the difference of computation results from the correct
results. The correct resutls can either be read from some csv file or we
can compute them using the simple CPU alg.