Add skeletons for optimized naive implementations
Add skeletons for three separate optimizations of naive implementation.

First optimization changes the matrix we parallelize the computation
over from the result matrix to the reference matrix. This way, each
thread block loads a piece of the reference matrix and computes
everything there is to be computed using this piece of reference matrix.
This tries to reduce the amount of data moved to/from global memory.

Second is to optimize the original implementation. Each cycle in the
original implementation rereads 15/16 of the data used in the previous
cycle, so prefetching a chunk of data into shared memory and reading it
from there should speed this up.

First try is to change this to work with matrix rows. We load a part of
a row from the reference matrix. We then have two piece ring buffer
containing data from deformed matrix, which is twice as large as the
reference matrix row buffer.

We first preread one part of the deformed buffer. Then in each cycle, we
read the reference buffer and top part of the deformed buffer, where
each thread reads a single value and stores it in memory. THis should
lead to best possible data throughput.

The threads then go through the reference buffer and use the
corresponding value from deformed buffer. They go through at most the
size of reference buffer. As they start in the lower part of the
deformed buffer and they do at most size of the reference buffer steps,
they should not overrun and they should all read up to the upper part of
the deformed buffer. We then continue the cycle, reading the lower part
and using the properties of ring buffer to continue reading from it.

Third optimizations is similar as the one with row ring buffer, but
instead using a submatrix ring buffer.

Only the row ring buffer somewhat works, and only for small matricies.
There is a bug somewhere in there that causes the threads to do an out
of bound access to the shared memory buffers.

Fixed problem with shared memory, where the extern declaration was being
instantiated for each datatype, leading to compiler complaining that we
are redefining the variable with different types. Now we have a wrapper
that declares it as double2 for 16B alignment and then reinterpret casts
it into the required type, preventing redeclaration.