Refactored the row caching algorithm
Refactored and fixed the row caching algorithm, now it is more object
oriented and readable. Ring buffers are now a class instead of a bunch
of code in the method.

Both ref and def need to be ring buffers, as the most negative shifts
broadcast from ref and need def to be ring buffer and the most positive
shifts broadcast from def and need ref to be ring buffer.

Added class for handling loading to shared memory, for representing
slice of a matrix, slice of a row etc.

Increased shared memory for the second ring buffer.

Fixed switched x and y in the input generator command line input.
Separated them into columns and rows CLI input variables, as before it
was not really descriptive.

Changed the classes representing each algorithm, so that they follow
best practices and allow us to by default measure the total runtime of
the algorithm. Basically removed most of the public abstract methods,
and replaced them with a pair of public non-virtual method, preserving
the parent class' control of the entrypoint, and protected abstract
method, forcing subclass to provide implementation.

Added commandline argument selecting the algorithm to be used. We no
longer need to recompile to test or measure different algorithms.

Fixed validation which would not work if the valid value was 0. So now
it computes mean difference without any scaling.