Removed data_single
data_single was basically just a subset of functionality of data_array,
so now what would have been data_single is just data_array with
num_matrices 1.

Add default arguments to data_array methods so that they are simple to
use for data arrays with 1 matrix.

Change data array csv format. Instead of having two rows of header data,
we now use single row and comment it based on what numpy does by default
with a header. Header is now in the format "#
<rows>,<cols>,<num_matrices>" without the quotes. We will try to
tolerate missing num_matrices and assume 1 matrix when missing, but all
outputs will have the full header even for data array with a single
matrix.

Change input generator to generate data_arrays.
Add generation of multi-matrix data arrays to input generator.

The change to data_arrays everywhere greatly simplifies the inheritance
between different algorithm types. As they all have the same data type
for inputs and results, we can now expose these in the root
cross_corr_alg class. We could even move the data there, but in my
opinion it is better to keep data members in the leaf classes and just
require methods by declaring abstract methods in the parents.

Remove the implementations for multi-file inputs. We may return them in
the future, but for now it will simlify the implementation of additional
algorithms if we do not need to care about it. The loading and storing
implementations are left in the code, we just don't require the
algorithms to use them in any implementation for now.

Add implementation for different algorithms in the matlab validator.
This allows us to precompute valid results for different algorithms and
validate our C++ implementation against them. This may also be the
implementation we benchmark against in the future.