Add n_to_mn version of original algorithm
Add n_to_mn version of original algorithm. By itself it is designed for
n_to_mn processing, but up until now we have been using just a limited
portion of the capability for one_to_one and one_to_many algorithms.

Add n_to_mn cpu computation of valid output to C++ source.

Add n_to_mn to the matlab implementation.

Fix matlab not exiting on error and leaving us in the matlab "shell".
Now we catch the exception, print it as before and then just exit.

Move the CUDA implementation of each algorithm to separate .cu file, so
that it is better organized and easier to find.