Add support for neste CPU and GPU measurements
INstead of having single start time and single pair of cuda events
stored in the stopwatch class, we instead for each measurement allocate
an instance of separate class, either cpu_measurement or
cuda_measurement. These contain the label and either the start time or
the pair of cuda events.

To prevent unnecessary overhead during measurement, we preallocate pairs
of CUDA events for all possible measurements, which we are given as
argument in the stopwatch constructor. This way we don't have to do any
allocations and deallocations during the measurement itself.

This may lead to unnecessarily too many CUDA events allocated, but that
should not be a problem, hopefully.