Get fft_one_one working, add fft validation
Add -lineinfo compilation switch for Nsight Compute profiling, so that
it can match the SASS/PTX code to source code lines.

Add loading of multiple matrices from a single file when loading vector
of files. We might use it to load all inputs into a single buffer,
transfer that buffer to GPU and run all the FFTs at once on all the
inputs. Then the different types of algorithms just depend on the
hadamard we do on the results of FFT. Add number of matrices loaded from
each file.

Fix validation. We were passing the arguments to compare_results in a
wrong order, it should have been result first, then valid data. It did
not really matter until we started using FFT, where the result is
normalized and the valid data is not.

Remove compare_results method. Instead we do the result normalization
straight in the algorithm class, which is the place we should know about
the existence of FFT algorithm at all. The validation code then receives
just two data arrays which it compares, so now the order of the
arguments does not really matter, just to the sign of the mean.

Remove is_fft argument from validation command. As explained above,
validation should just compare two input matrices and ignore how they
were created. Any normalization will have to be done during the
computation itself or separately.

Fix the csv header when writing in C++. We were missing the "# " prefix
to the header.

Change the fft_one_one implementation so that it uses cuFFTPlanMany to
run the FFT transform of ref and target at once. To be able to do this,
we upload both the ref and target inputs into a single device buffer.