Add fft implementations
Add fft implementations for one_to_many, n_to_mn and n_to_m. one_to_many
and n_to_mn can use the original hadamard implementation and reuse most
of the code from the one_to_one algorithm.

The n_to_m has to be implemented differently, as we cannot use the
original hadamard kernel as it is just not designed for this.
We cannot overwrite the target matrices, so we have to allocate
additional hadamard_results buffer for results of the hadamard
multiplication.

Standardize algorithm naming, everything is now *_to_*, even the
one_to_one and one_to_many.

Add new hadamard kernel designed for n_to_m. We add two different
implementations, one which uses data paralellism of the output values
and one which uses paralellism of the right matrices. The output one
should lead to more parallelism, but cannot use any explicit caching and
data reuse. We can test this further in the future.