Fix CUDA time measurement
Copy-pasting is a frequent source of errors. The get%stop method was
returning the start_ event, so any measurement was measuring time
between start and start, which is of course 0.

Add missing CUCH macros around CUDA calls in the stopwatch
implementation. We were ignoring any CUDA errors up until now.

Move the measurement csv output before validation, so that if validation
takes some time, we can already start eyeball processing measurements.

Switch measurments of cuFFT from cuda to CPU, as we don't know how much
stuff the library does on the CPU, so we want to measure the whole
runtime of the funciton call, not just the duration between first and
last kernel calls inside.