Add benchmarking scripts
Add script simplifies benchmarking the algorithms in the C++
implementation with multiple input sizes.

Move input generator under the benchmarking scripts.

Change C++ algorithm output so that it can be easily disabled. THis way
the only thing the C++ program prints is the stats of the output
compared to valid output if validation is enabled, or nothing if
validation is disabled.

ADd validate_input_size to the C++ implementation and to all algorithms,
so that we can check if the given input size can be used with given
algorithm.

Change result stats output so that it can be printed in CSV format,
either with or without header. This enables us to append these results
from consecutive runs into a single CSV file.

Add benchmarking scripts. They implement benchmarking, validation, valid
output generation and input generation.
The benchmarking script takes list of algorithms of the same type and a
list of input sizes usable with the algorithms of the given type. THen
it runs each algorithm for each size of input, running it for the given
amount of iterations. Inputs of each size are generated before running
algorithms for the given size. It can optionally validate the output of
each iteration by precomputing the valid output for the generated
inputs.