Add scipy cross-correlation implementation
Add Python CPU-only implementation of cross-correlation as provided by
scipy. This should serve as benchmark of the slowest possible used
implementation compared to the GPU implementations, even the naive
algorithm one.

Add clarification of the ordering of the outputs in n_to_mn algorithms.
We first want the results of the cross-correlation of the n matrices
from the left input with the corresponding matrix from the first n
matrices of the right input, then following should be the result of
cross-correlating the n matrices from the left input with the
corresponding matrix from the matrices [n,2n) from the right input etc.
ending with the cross-correlation of the n matrices from the left input
with the corresponding matrix from the last n matrices from the right
input, i.e. matrices [(m-1)*n,m*n).