Enable the use of other existing cross-corr implementation as validators
Move matlab implementation under existing, as it is not anything special
and can be used as any other existing implementation to compute the
validation data.

Remove the unnecessary valida.m and run_validation.sh scripts which were
used previously for validation but are currently unused.

Update the default validator path in the benchmarking scripts.
We still default to matlab.

Add compute_valid_results.sh script to the SciPy implementation so that
it can be used as validator instead of the matlab implementation. Both
these scripts present the same API and can thus be used interchangably.