Add external args to algorithms, upgrade benchmarking
Add external arguments to algorithms. We now allow user to provide a
path to a JSON file which is then loaded and given to the algorithm
constructor to read whatever it needs from it.

Add nlohmann json for parsing JSON in C++.

Change how we specify the benchmarks to be ran by the benchmarking
scripts. Instead of providing everythin on the commandline, we now
specify everything in a .yml file which one of the few arguments given
to the benchmarking script. The .yml file contains benchmarking groups,
which share the same input data. These should be used to compare
similiar algorithms to each other.

The .yml file also allows us to provide the algorithm specific arguments
for each of the algorithsm, as mentioned above.

The yml file contains a single global configuration under
benchmark/config, which is inherited by all groups. Each group can then
override any of the settings in the global configuration.

This is to reduce the amount of data we need to input to run simple
benchmarks.