Change measurements, change validation, add algorithm
Change C++ measurements.
Split the prepare method into load, prepare and transfer methods. Load
will just load the data from disk, prepare allocates everything and
transfer does data transfer to device. All of these steps, together with
run and finalize, are now measured by the root cross_corr_alg class and
do not have to be explicitly measured by child classes.
The labels for these root measurements are defined by the root
cross_corr_alg class, and are prepended to the labels returned by the
abstract method measurement_labels_impl implemented by child classes.
For now the implementation expects only the root and the leaf classes to
add labels and do measurements, but this could possible be upgraded in
the future.

Labels in leaf classes have to be passed to the measurement macros after
they are processed by the label_index method from the root class, which
adds the number of root labels to the index so that the resulting index
points to the leaf labels appended after the root labels.

Change validation. Instead of going with straight difference of
corresponding result values, we now do relative difference, where the
absolute difference value is divided by the larger of the two values.
This should allow simpler detection of errors regardless of the size of
the outputs. Before the larger input sizes resulted in bigger output
values and proportionally bigger errors, which were hard to filter. NOw
regardless of input size, the error values should be the same.

Add fft_reduced_transfer algorithm. Whereas the original algorithm adds
the padding to the input data in host memory and then transfers this 4x
enlarged data to the device memory, this new algorithm transfers just
the input data and then uses a new scatter kernel to scatter the input
matrices into the correct parts of a zeroed out buffer for the padded
matrices.

Add optional normalization to the validate command of the C++ program.
This allows us to validate even denormalized outputs of fft algorihtms.

Improve benchmarking progress output. Change default run names to use
algorithm name instead of just the index. Add verbose mode which prints
the executed command so that we can debug it if the command fails.

Improve benchmarking arguments. Add the ability to run just a selection
of groups from the benchmarking definition by providing an optional list
of group names after the path to the benchmarking definition on the
commandline.

Add data type to validators. Instead of running with double, as matlab
and Scipy do by default, we now can specify the data type on the
commandline as "single" or "double" for single or double floating point.

Add validation of result stats, which goes through the statistics of
validation and checks that the maximum error is below a specified
threshold. We can add additional optional checks. This command allows a
list of output_stats.csv files to be checked, and optional list of
groups from which it will pull all the output_stats.csv files and check
those.

Add first part of benchmark result visualization. For now the we have
visualization of total time taken. But with implementation of result
loading etc. for future visualization.