Add data type switch, add input subarray transformation
Add data type switch to the C++ cross-correlation. Instead of having to
recompile to switch from using float or double, we now compile for both
and can select using a new commandline option data_type. This option
defaults to single.

Matlab and Scipy seem to be using double for the cross-correlation
computation, as when we run with single, there is arround 0.16 error, so
16% of the value size off. WHen we run with double, we go down to 1e-10
error, so basically no error. In a future commit, we will have to fix
matlab and scipy to really use the data type we provide when computing
the validation results.

Fix one_to_one fft reduced transfer algorithm, which was using ref_
matrix_size in places where padded matrix size was supposed to be used.
These were leftovers from copy pasting the original fft algorithm, which
has ref_ matrix padded.

Add first input array transformation, subarray extraction, to input
generator. This tool allows us to extract single or multiple matrices
from matrices tored in a single input csv file. This is useful for
debugging the more complex algorithms such as n_to_mn and n_to_m, where
we can detect the combination producing the invalid result and extract
the two source matrices and the resulting matrix to validate against.