Add Wall, Wextra, Wpedantic and fix the warnings
Add compiler switches to warn us about everything it can. When compiling
with nvcc, we need to explicitly pass these to the underlying compiler,
which is done in the CMakeLists.txt.

Add missing cufft error switch cases.

Mark any parameters which are fine if they are not used as maybe_unused.
These are mainly the algorithm constructors to preserve signature
compatibility for usage in main.cpp.

Fix singed/unsigned comparison which mostly happened in for loops with
iteration variable declared using auto, which deduced its type to int
and we were comparing it to dsize_t which is unsigned int.
Now we explicitly type the iteration variables as dsize_t.

Add try catches with finer granularity around each of the algorithm
steps so that we immediately know where an error came from. This can be
useful if we ever make a mistake in an implementation of any of the
algorithms.

Add implementation of fft_reduced_transfer for one_to_many.