Add support for switching data type used for computation to benchmarking
We can now specify if the computation by both the C++/CUDA, matlab and
scipy implementations should be done in single or double floating point
data types. This allows us to measure both the differences in speed and
percision between the same algorithms using different data types and the
relative changes in behavior when compared to the changes of other
algorihtms and technologies.

This was prompted by the instability of results of the C++/CUDA
implementation when computed in the 32bit floating point, where it has a
relatively huge error compared to matlab/scipy, whereas it is
practically equal with matlab/scipy when all are computed in double.

This will need further investigation.

The data_type entry was added to the global config and to the group
config, so that each group can use different data type and we can
compare the times and results between them.

The same type is used both for computation and for generation of
validation data, to see the error in the C++/CUDA implementations.