Add algorithm arguments to benchmarking
Algorithm arguments, such as number of threads per block or number of
items per thread can now be defined per algorithm and are passed to the
algorithm constructor as JSON object. Any algorithm can then read
whatever data it needs from the JSON object.

Add argument properties to the time measurements csv output. The
arguments read from JSON can now be added to the CSV file containing
timing results, so that we can do easier visualization and processing
later. The additional_properties as we call them are exposed by each
algorithm through the additional_properties method, which returns a
vector of string pairs, each pair containing CSV label and value.
The data and optionally (when not appending) the labels are appended to
the time measurement CSV to each line, so we can use it later in
processing.

Add fft_reduced_transfer to n_to_mn.

Fix bug in one_to_many fft_reduced_transfer, which was allocating
results_ array too small due to missing argument to the data_array
constructor.

Change how we output CSV. Instead of each csv method expecting to be the
only data appended to the line, the methods now just append the data to
the current line without leading comma and trailing newline, so we can
chain them together and add the comma or newline explicitly outside the
method.

Add argument parsing and generation of argument value combinations to
benchmarking scripts. The benchmarking definition can now contain
argument specification, which is processed by the benchmarking script.
All lists in the arguments for given algorithm are gathered and all
combinations of the values in these lists are created. Instead of
generating a single Run per run declaration, we now generate as many as
there are combinations of their arguments. The argument values are then
appended to the name of the run to make it unique.

Arguments which are not lists or are empty lists are left as is and are
passed to the algorithm together with each combination of list
arguments. Lists of a single value are just extracted and passed on as
single value. This allows us to pass a list of values as a single
argument, just by enclosing the list of values in another list.
Combinations are not generated recursively.

Change output naming conventions in the benchmarking scripts. Now
separate pieces of a name are separated by -, and subsections of a
single part of the name are separated by _. So run index, input index,
run name and input name are separated by -, parts of the run name are
separated by _, same as parts of the input name.