Change cuda time measurement
INstead of synchronizing the device to measure the event time
immediately, we instead just insert the events and record that the event
pair was used. Then in the finalization step as the last thing we go
through the events and read the time from the used event pairs.

This should reduce the overhead during algorithm runtime and move it
outside the measured part of the code.

Change naming convention of the benchmarking results. The Run name now
encloses the argument derived part in a pair of underscores from both
sides, so that it is easily parsable using regex in visalization.

Update visualization to reflect the changes to benchamrking naming
conventions.