Add external runs to benchmarking
Add external runs to benchmarking. The previously benchmarked runs of
C++/CUDA implementation are now called internal, as they are internal to
our project. The external implementations implemnet a simple API, where
they present a single executable with 7 parameters, algorithm, data
type, number of iterations, input1, input2, output and timings path.
These are then used to run the required number of iterations, recording
the same base timings as the internal implementation and writing these
to the timings path.

The external runs in the benchmark yaml specification are distinguished
by the type element of the run object. If no type or type of value
"internal" is given, intenral Run is parsed. If type with value
"external" is present, ExternlRun class is used and parsed.

Move some arguments from the run_benchmarks method of InternalRun into
the constructor and store them in the instance. This is done to preserve
the bare minimum of arguments which are shared between Internal and
external runs.

Move args file generation for the internal algorithms to the abstract
prepare method. THe args file path is now created during internal Run
instance creation and is used in prepare.

Add base_dir_path to global config, which is the path to the directory
containing the benchmark yaml definition. This path is used as base for
any relative paths used in the benchmark definition file, such as paths
to the external executable scripts for external runs.

Improve executable validate data, which can now take a list of target
data to be validated. It then does all the validations in a single call.

Improve executable validate data which can now output csv output, the
same as executable run validation. THis is used to validate outputs of
the external runs.

Fix matlab scripts using deprecated -r option to run the matlab scripts.
We instead use the suggested -batch option, which disables the matlab
header message and by default runs with no display, no splash and no
desktop. The options disabling these are left in anyway as we really
don't want those things.

Add time measurement to matlab implementation. We use the tic, tac
function pair, with preallocated array to store the measured time in the
lowest possible overhead. The times measured are the same as measured by
the base corss_corr_alg class in C++/CUDA implementation.

Improve matlab algorithms to all preread the whole input at once and
then use indexing into the input. THis allows us to move matlab input
reads out of the alg_type switch and should make the loading faster and
easier to measure.

Improve matlab result movements, consolidating it first in gpu memory
and gathering it to host memory once after the whole computation.
Instead of gathering each result and pasting it into the correct part of
the result_matrix, we now preallocate the g_result matrix, i.e. result
matrix on the gpu and paste the result to the correct part of this
matrix, using the same indexing as with the previous gather on host.
Only after the whole computation do we gather the result to host memory.
This simplifies measurement and should make the implementation faster.

Add timing output to matlab cross correlation. If the optional
timings_path variable is set, we transform the the timings into a table
and write it to csv. Thankfully matlab by itself does appending well,
priting header only if the output file is empty.

Add timings to python implementation. The timings use
time.perf_counter_ns() to hopefully use the clock with the best
resolution available to measure in nanoseconds, the same as C++/CUDA
implementation. This is then saved to CSV if the optional timings_path
is provided, where header is only written if the file does not exist.
This is inconsistent with C++/CUDA where the header is explicitly
controlled by the --append flag and with matlab which does the best
things and adds header only if there is no header. We may need to fix
this in the future.

Add two versions of matlab benchmark script. One does the iterations
inside matlab, which leads to matlab caching most if not all of the
computation. Seconds does the iteration in bash, starting matlab with
each iteration. This leads to much slower execution times, but may be
more representative of the usecase we target.

Add python benchmark script, which just runs the required number of
iterations of the python cross corr script with additional timings_path
argument.