Fix shuffle implementation
The shuffle implementation is now functioning. I just made it
unnecessarily too complicated, it really only needed to properly compute
the warp_x_right_start and warp_x_right_end.

Now we always go from left to right over the matrix, broadcasting from
the right matrix and shuffling the left matrix. If broadcast value does
not overlap in the shift computed by the current thread, the thread will
load 0 as it should access the left matrix out of bounds and thus is
ignored. With shuffling, it will receive the correct values the moment
the computation reaches it's shift.

Remove debug prints from nainve_shuffle.cu.

Add profiling of the current implementations of original and shuffle
algorithsm with the 64x64x1x1 inputs. With this size of matrix, there is
a problem in occupancy which I will try to solve in future commits.

WARNING: Do not profile or benchmark with -G and other flags in
CMakeLists.txt. It makes CUDA code abuse local memory and slows it down.
Always check that we are compiling at most with -lineinfo for the
profiler.